Some small refactoring of collision code
I cheated a little bit and glanced at the Rust version of the
Ray Tracing in One Weekend book and confirmed what I had suspected,
passing in a reference to HitRecord was not a great solution. The hit()
functions should return an Option<HitRecord> which lets the caller get
the data _and_ check if it is None.